% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoplot.R
\name{autoplot.DynamicStrategies}
\alias{autoplot.DynamicStrategies}
\title{Easy Inspection of a Dynamic Strategy Simulation}
\usage{
\method{autoplot}{DynamicStrategies}(object, ...)
}
\arguments{
\item{object}{An object of the `DynamicStrategies` class.}

\item{...}{Additional arguments to be passed to `autoplot`.}
}
\value{
A \code{ggplot2} object.
}
\description{
Extends the `autoplot` method for the `DynamicStrategies` class.
}
\examples{
library(ggplot2)

simulate_strategy() |> autoplot()
}
