% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_stats.R
\name{extract_stats}
\alias{extract_stats}
\title{Extract Portfolio Statistics from a Dynamic Simulation}
\usage{
extract_stats(simulation, level = 0.01)
}
\arguments{
\item{simulation}{An object of the \code{DynamicStrategies} class.}

\item{level}{A number with the desired probability level. The default is
\code{level = 0.01}.}
}
\value{
A tibble with 3 columns.
}
\description{
Computes the mean, standard deviation, skewness, kurtosis, Value-at-Risk (VaR)
and Conditional Value-at-Risk CVaR) of a strategy.
}
\examples{
utility <- simulate_strategy(strategy = "max_utility")
cppi    <- simulate_strategy(strategy = "cppi")
extract_stats(utility)
extract_stats(cppi)
}
