% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_sensivities.R
\name{plot_sensivities}
\alias{plot_sensivities}
\title{Stress-Test a Dynamic Strategy}
\usage{
plot_sensivities(strategy, variable, from, to, size, ...)
}
\arguments{
\item{strategy}{An object of the \code{DynamicStrategies} class.}

\item{variable}{A \code{character} with the parameter to loop-over.}

\item{from}{An \code{integer} with the initial parameter value.}

\item{to}{An \code{integer} with the final parameter value.}

\item{size}{A \code{integer} with the total number of iterations.}

\item{...}{Currently not used.}
}
\value{
A \code{ggplot2} object.
}
\description{
This function loops over one of the arguments in \code{\link{simulate_strategy}}
to make explicit the impact of that variable in the desired allocation policy.
}
\examples{
# stress-test the `max_utility` strategy by changing the `allocation` parameter
x <- simulate_strategy()
plot_sensivities(strategy = x, variable = "allocation", from = 0.0, to = 1, size = 20)

# stress-test the `cppi` strategy by changing the `floor` parameter
y <- simulate_strategy(strategy = "cppi")
plot_sensivities(strategy = y, variable = "floor", from = 7000, to = 10000, size = 20)

# stress-test on the `obpi` strategy by changing the risk-free rate
y <- simulate_strategy(strategy = "obpi")
plot_sensivities(strategy = y, variable = "rf", from = 0, to = 0.2, size = 20)
}
