% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_strategy.R
\name{simulate_strategy}
\alias{simulate_strategy}
\title{Convex vs Concave Portfolio Simulation}
\usage{
simulate_strategy(
  budget = 10000,
  horizon = 6/12,
  step = 1/252,
  mu = 0.2,
  sigma = 0.4,
  rf = 0.04,
  n_simul = 10000,
  strategy = c("max_utility", "buy_hold", "cppi", "obpi"),
  allocation = 0.5,
  floor = budget * 0.9,
  multiple = 10,
  aggressiveness = 0.5
)
}
\arguments{
\item{budget}{A \code{double}. The initial investment.}

\item{horizon}{A \code{double}. The investment horizon, in years.}

\item{step}{A \code{double}. The time frequency, in years.}

\item{mu}{A \code{double}. The mean of the process, in years.}

\item{sigma}{A \code{double}. The volatility of the process, in years.}

\item{rf}{A \code{double}. The risk-free rate, in years.}

\item{n_simul}{A \code{integer}. The number of simulations to be conducted.}

\item{strategy}{A \code{character}. One of: "max_utility", "buy_hold", "cppi", "obpi".}

\item{allocation}{A \code{double} between \code{0} and \code{1}. Only used for
"max_utility" and "buy_hold".}

\item{floor}{A \code{double}. The amount the investor doesn't want to loose. Only
used if \code{strategy = "cppi"}.}

\item{multiple}{A \code{integer} showing how aggressive the investor should be.}

\item{aggressiveness}{A \code{double} between \code{0} and \code{1} showing how
aggressive the investor should be. Higher numbers are connected with higher convexity.}
}
\value{
A tidy \code{tibble} with 6 rows and 2 columns.
}
\description{
This function simulates the payoff of different kinds of dynamic strategies.
Depending on the option chosen, the output may be concave or convex.
}
\examples{
simulate_strategy()
}
